modified:   backend/predict.py 	modified:   src/views/MainView.vue

+++ backend/predict.py
@@ -33,7 +33,7 @@ class Predictor(BasePredictor):
         self.model = WhisperModel(
             model_name,
             device="cuda" if torch.cuda.is_available() else "cpu",
-            compute_type="float32",
+            compute_type="float16",
         )
         hf_auth_token = os.getenv("HF_TOKEN")
         if not hf_auth_token:
